refactor(seth): split out cli utils
